AI_ScriptGen_Karanji/requirements.txt
 flask
python-docx
python-pptx
pdfminer.six
openai
python-dotenv
